Rename modules to convey private intent
