+++ openchemistry/__init__.py
@@ -1,3 +1,3 @@
-from .utils import hash_object, camel_to_space, parse_image_name
+from ._utils import hash_object, camel_to_space, parse_image_name
 from .io import Psi4Reader, NWChemJsonReader, CjsonReader
-from .public_api import load, find_structure, find_calculation, find_molecule, monitor, queue, find_spectra
+from .api import load, find_structure, find_calculation, find_molecule, monitor, queue, find_spectra
