+++ openchemistry/_notebook.py
